requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
dataclasses-json==0.6.7
Django==5.0.7
django-htmx==1.18.0
django-markdownify==0.9.5
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
hashids==1.3.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.18
langchain-google-genai==1.0.7
langchain-text-splitters==0.2.2
langsmith==0.1.85
Markdown==3.6
markdownify==0.12.1
marshmallow==3.21.3
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.6
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
rsa==4.9
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.0
tenacity==8.5.0
tinycss2==1.2.1
tqdm==4.66.4
Twisted==24.3.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
webencodings==0.5.1
yarl==1.9.4
zope.interface==6.4.post2
